Mudando os id's das interfaces para long e tornando o retorno dos metodos de get e getbyid do service genéricos

+++ Snacker.Domain/Interfaces/IBaseRepository.cs
@@ -9,10 +9,10 @@ namespace Snacker.Domain.Interfaces
 
         void Update(TEntity obj);
 
-        void Delete(int id);
+        void Delete(long id);
 
         ICollection<TEntity> Select();
 
-        TEntity Select(int id);
+        TEntity Select(long id);
     }
 }
